rest-framework-auth-jwt/requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==44.0.3
Django==5.2.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
parameterized==0.9.0
pillow==11.2.1
pycparser==2.22
PyJWT==2.9.0
sqlparse==0.5.3
